Reviews Model Done (9.1)

+++ backend/controllers/bootcamps.js
@@ -33,6 +33,19 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @route /api/v1/bootcamps/
 // @access Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+	// Add user to req.body
+	req.body.user = req.user.id;
+
+	// Check for published bootcamp 
+	const publishedBootcamps = await Bootcamp.findOne({ user : req.user.id });
+
+	// If the user is not an admin, they can only add one bootcamp
+	if(publishedBootcamps && req.user.role !== 'admin') {
+		return next(new ErrorResponse(
+			`The user with ID ${req.user.id} has already published a bootcamp`,
+			400
+		));
+	}
 	const bootcamp = await Bootcamp.create(req.body);
 	res.status(201).json({
 		success : true, 
@@ -45,20 +58,27 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
 	const id  = req.params.id;
 	const data = req.body;
-	const bootcamp = await Bootcamp.findByIdAndUpdate(id, data, {
-		new: true,
-		runValidators: true
-	});
+	let bootcamp = await Bootcamp.findById(id);
 
 	if(!bootcamp){
-		return next(
-			new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
-		)
+		return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404));
 	}
+
+	//Make sure that the user is bootcamp owner
+	if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+		return next(
+			new ErrorResponse(`User ${req.user.id} is not authorized to update this bootcamp.`, 401)
+		);	
+	} 
+
+	bootcamp = await Bootcamp.findOneAndUpdate(bootcamp, req.body, {
+		new : true,
+		runvalidators : true,
+	});
 
 	res.status(200).json({
 		success : true, 
-		data: bootcamp
+		data : bootcamp
 	});
 });
 // @desc Get bootcamps within a radius
@@ -78,12 +98,12 @@ exports.getBootcampsInRadius = asyncHandler(async (req, res, next) => {
 
 	const bootcamps = await Bootcamp.find({
 		location: { $geoWithin: { $centerSphere: [ [ lng, lat ], radius ] } }
-	})
+	});
 	res.status(200).json({
 		success: true,
 		count:  bootcamps.length,
 		data: bootcamps
-	})
+	});
 });
 // @desc delete a bootcamp using id
 // @route DELETE /api/v1/bootcamps/:id
@@ -99,10 +119,17 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
 		)
 	}
 
+	//Make sure that the user is bootcamp owner
+	if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+		return next(
+			new ErrorResponse(`User ${req.user.id} is not authorized to delete this bootcamp.`, 401)
+		);	
+	} 
+
 	await bootcamp.deleteOne()
 	await Course.deleteMany( { bootcamp: bootcamp._id });
 
-	res.status(200).json({success : true, data: {} });
+	res.status(200).json({ success : true, data: {} });
 });
 
 // @desc Upload photo for bootcamp
@@ -117,6 +144,14 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
 			new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
 		)
 	}
+
+	//Make sure that the user is bootcamp owner
+	if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+		return next(
+			new ErrorResponse(`User ${req.user.id} is not authorized to update this bootcamp.`, 401)
+		);	
+	} 
+	
 	if(!req.files){
 		return next(new ErrorResponse(`Please upload a file`, 400));
 	}
