+++ backend/controllers/courses.js
@@ -4,7 +4,9 @@ const { parse } = require('dotenv');
 const { findById } = require('../models/Bootcamp');
 const Bootcamp = require('../models/Bootcamp');
 const Course = require('../models/Course');
+
 // @desc Get all Courses
+// @route GET /api/v1/courses
 // @route GET /api/v1/bootcamps/:bootcampId/courses
 // @access Public
 exports.getCourses = asyncHandler(async (req, res, next) => {
@@ -32,7 +34,7 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 	});
 
 	if(!course){
-		return next(new ErrorResponse(`No course with the id of ${req.params.id}`, 404))
+		return next(new ErrorResponse(`No course found with the id of ${req.params.id}`, 404))
 	}
 
 	res.status(200).json({
@@ -46,14 +48,21 @@ exports.getCourse = asyncHandler(async (req, res, next) => {
 // @access Private
 exports.addCourse = asyncHandler(async (req, res, next) => {
 	req.body.bootcamp = req.params.bootcampId;
-
+	req.body.user = req.user.id;
 
 	const bootcamp = await Bootcamp.findById(req.params.bootcampId);
 
 	if(!bootcamp){
-		return next(new ErrorResponse(`No bootcamp with the id of ${req.params.bootcampId}`, 404))
+		return next(new ErrorResponse(`No bootcamp found with the id of ${req.params.bootcampId}`, 404));
 	}
 
+	//Make sure that the user is bootcamp owner
+	if(bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+		return next(
+			new ErrorResponse(`User ${req.user.name} is not authorized to add a course to bootcamp ${bootcamp.name}.`, 401)
+		);	
+	} 
+	
 	const course = await Course.create(req.body);
 
 	res.status(200).json({
@@ -67,10 +76,17 @@ exports.addCourse = asyncHandler(async (req, res, next) => {
 // @access Private
 exports.updateCourse = asyncHandler(async (req, res, next) => {
 	let course = await Course.findById(req.params.id);
-
+	
 	if(!course){
-		return next(new ErrorResponse(`No course with the id of ${req.params.bootcampId}`, 404))
+		return next(new ErrorResponse(`No course found with the id of ${req.params.bootcampId}`, 404))
 	}
+
+	//Make sure that the user is course owner
+	if(course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+		return next(
+			new ErrorResponse(`User ${req.user.name} is not authorized to update course ${course.title}.`, 401)
+		);	
+	} 
 
 	course = await Course.findByIdAndUpdate(req.params.id, req.body, {
 		new: true,
@@ -91,7 +107,7 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 
 		if(!bootcamp){
 			return next(
-				new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+				new ErrorResponse(`Course not found with id of ${req.params.id}`, 404)
 			)
 		}
 
@@ -140,13 +156,22 @@ exports.updateBootcamp = asyncHandler(async (req, res, next) => {
 // @access Private
 exports.deleteCourse = asyncHandler(async (req, res, next) => {
 		const id = req.params.id;
-		const course = await Course.findByIdAndDelete(id);
+		let course = await Course.findById(id);
 		
 		if(!course){
 			return next(
-				new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
+				new ErrorResponse(`Course not found with id of ${id}`, 404)
 			)
 		}
+		
+		//Make sure that the user is course owner
+		if(course.user.toString() !== req.user.id && req.user.role !== 'admin') {
+			return next(
+				new ErrorResponse(`User ${req.user.name} is not authorized to delete course ${course.title}.`, 401)
+			);	
+		} 
+		
+		course = await Course.findByIdAndDelete(id);
 
 		res.status(200).json({success : true, data: {} });
 });
